Revert "Guest checkout breaks user account with the same email (#732)"
This reverts commit f740b834ab4f5123c307694e685dac963f8e5d3b.
