--- phoenix-scala/gatling/src/main/scala/playground/GatlingSimulations.scala
@@ -1,9 +1,0 @@
-package playground
-
-import seeds.Seeder
-
-// running gatling sims as an integration tests
-// https://github.com/FoxComm/highlander/pull/732#discussion_r98843060
-object GatlingSimulations extends App with Seeder {
-  runSimulation[EvilGuestSimulationLocally]()
-}
