+++ phoenix-scala/gatling/src/main/scala/seeds/requests/Cart.scala
@@ -11,7 +11,7 @@ import payloads.LineItemPayloads.UpdateLineItemsPayload
 import payloads.CartPayloads.CreateCart
 import payloads.OrderPayloads.OrderTimeMachine
 import payloads.UpdateShippingMethod
-import helpers.dbFeeder
+import seeds.dbFeeder
 import seeds.requests.Auth._
 import seeds.requests.Payments._
