+++ phoenix-scala/gatling/src/main/scala/seeds/simulations/CustomerActivitySimulation.scala
@@ -7,7 +7,6 @@ import seeds.requests.Addresses._
 import seeds.requests.Auth._
 import seeds.requests.Cart._
 import seeds.requests.Customers._
-import helpers._
 
 class CustomerActivitySimulation extends Simulation {
