+++ phoenix-scala/gatling/src/main/scala/seeds/simulations/OneshotSeedsSimulation.scala
@@ -3,7 +3,7 @@ package seeds.simulations
 import io.gatling.core.Predef._
 import io.gatling.core.scenario.Simulation
 import seeds.Conf._
-import helpers._
+import seeds._
 import seeds.requests.Auth._
 
 class OneshotSeedsSimulation extends Simulation {
