+++ phoenix-scala/project/Tasks.scala
@@ -14,5 +14,4 @@ object Tasks {
   lazy val seedOneshot    = inputKey[Unit]("Run oneshot gatling seeds")
   lazy val seedContinuous = inputKey[Unit]("Run continuous gatling seeds")
 
-  lazy val testSimulations = taskKey[Unit]("Run gatling test simulations")
 }
